+++ packages/braintree-integration/src/braintree-fastlane/braintree-fastlane-customer-strategy.ts
@@ -78,13 +78,8 @@ export default class BraintreeFastlaneCustomerStrategy implements CustomerStrate
 
         const state = this.paymentIntegrationService.getState();
         const customer = state.getCustomerOrThrow();
-        const features = state.getStoreConfigOrThrow().checkoutSettings.features;
-        const shouldSkipFastlaneForStoredMembers =
-            features &&
-            features['PAYPAL-4001.braintree_fastlane_stored_member_flow_removal'] &&
-            !customer.isGuest;
 
-        if (this.isAcceleratedCheckoutEnabled && !shouldSkipFastlaneForStoredMembers) {
+        if (this.isAcceleratedCheckoutEnabled && customer.isGuest) {
             const shouldRunAuthenticationFlow = await this.shouldRunAuthenticationFlow();
 
             if (
@@ -99,7 +94,10 @@ export default class BraintreeFastlaneCustomerStrategy implements CustomerStrate
             }
 
             if (shouldRunAuthenticationFlow && this.isFastlaneEnabled) {
-                await this.braintreeFastlaneUtils.runPayPalFastlaneAuthenticationFlowOrThrow();
+                await this.braintreeFastlaneUtils.runPayPalFastlaneAuthenticationFlowOrThrow(
+                    undefined,
+                    true,
+                );
             }
         }