+++ packages/core/src/payment/strategies/braintree/braintree-payment-processor.spec.ts
@@ -433,12 +433,15 @@ describe('BraintreePaymentProcessor', () => {
             expect(threeDSecureMock.verifyCard).toHaveBeenCalledWith({
                 addFrame: expect.any(Function),
                 removeFrame: expect.any(Function),
-                challengeRequested: true,
+                challengeRequested: expect.any(Boolean),
                 amount: 122,
                 bin: '123456',
                 nonce: 'tokenization_nonce',
                 onLookupComplete: expect.any(Function),
                 collectDeviceData: true,
+                additionalInformation: {
+                    acsWindowSize: '01',
+                },
             });
         });