+++ packages/core/src/payment/strategies/braintree/braintree.ts
@@ -416,6 +416,9 @@ export interface BraintreeThreeDSecureOptions {
     removeFrame(): void;
     onLookupComplete(data: BraintreeThreeDSecureVerificationData, next: () => void): void;
     collectDeviceData: boolean;
+    additionalInformation?: {
+        acsWindowSize?: '01' | '02' | '03' | '04' | '05';
+    };
 }
 
 interface BraintreeThreeDSecureVerificationData {
@@ -516,4 +519,5 @@ export interface BraintreeInitializationData {
     intent?: 'authorize' | 'order' | 'sale';
     isCreditEnabled?: boolean;
     isAcceleratedCheckoutEnabled?: boolean;
+    isFastlaneEnabled?: boolean; // TODO: remove this line when fastlane experiment will be rolled out to 100%
 }