+++ packages/paypal-commerce-integration/src/paypal-commerce-fastlane/paypal-commerce-fastlane-customer-strategy.ts
@@ -122,13 +122,8 @@ export default class PayPalCommerceFastlaneCustomerStrategy implements CustomerS
 
         const state = this.paymentIntegrationService.getState();
         const customer = state.getCustomerOrThrow();
-        const features = state.getStoreConfigOrThrow().checkoutSettings.features;
-        const shouldSkipFastlaneForStoredMembers =
-            features &&
-            features['PAYPAL-4001.paypal_commerce_fastlane_stored_member_flow_removal'] &&
-            !customer.isGuest;
 
-        if (this.isAcceleratedCheckoutFeatureEnabled && !shouldSkipFastlaneForStoredMembers) {
+        if (this.isAcceleratedCheckoutFeatureEnabled && customer.isGuest) {
             const shouldRunAuthenticationFlow = await this.shouldRunAuthenticationFlow();
 
             if (
@@ -232,9 +227,31 @@ export default class PayPalCommerceFastlaneCustomerStrategy implements CustomerS
             await this.paymentIntegrationService.updateBillingAddress(billingAddress);
         }
 
-        // Info: if not a digital item
         if (shippingAddress && cart.lineItems.physicalItems.length > 0) {
             await this.paymentIntegrationService.updateShippingAddress(shippingAddress);
+        }
+
+        const features = state.getStoreConfigOrThrow().checkoutSettings.features;
+        const shouldDisableFastlaneOneClickExperience =
+            features && features['PAYPAL-4142.disable_paypal_fastlane_one_click_experience'];
+
+        if (
+            shippingAddress &&
+            cart.lineItems.physicalItems.length > 0 &&
+            !shouldDisableFastlaneOneClickExperience
+        ) {
+            const consignments = state.getConsignments() || [];
+            const availableShippingOptions = consignments[0]?.availableShippingOptions || [];
+            const firstShippingOption = availableShippingOptions[0];
+            const recommendedShippingOption = availableShippingOptions.find(
+                (option) => option.isRecommended,
+            );
+
+            if (recommendedShippingOption || firstShippingOption) {
+                const shippingOptionId = recommendedShippingOption?.id || firstShippingOption.id;
+
+                await this.paymentIntegrationService.selectShippingOption(shippingOptionId);
+            }
         }
     }